Completed the layout of the game
I was able to display the tictactoe board in html.
Used flexbox display for that. Did some brainstorming
of the code in Javascript but decided it will be easier
to first make the code work, then try to refactor it
with factory and Constructor functions.

Next objective is to get the game to work. Once working
then I will refactor the code.