Refactor the codes with more functions
I was able to refactor my code so it can handle more
functions. Also I addded a new function that will handle
click event listenrs. So now the game is complete.
When I come back to this project, next will be to
refactor more, and style the UI. Also adding
a point system.